requirements.txt
streamlit
google-generativeai
langchain
chromadb
streamlit>=1.20
google-generativeai>=0.3.0
python-dotenv
fpdf
